pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eegfm-digest"
version = "0.1.0"
description = "Monthly EEG Foundation Model digest from arXiv with Gemini triage + summarization."
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "python-dateutil>=2.9.0",
  "PyYAML>=6.0.1",
  "pypdf>=4.2.0",
  "pdfminer.six>=20231228",
  "jsonschema>=4.22.0",
  "google-genai>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[project.scripts]
eegfm-digest = "eegfm_digest.run:main"
eegfm-batch = "eegfm_digest.batch:main"
eegfm-profile = "eegfm_digest.profile:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
